refactor(facets): use custom element for facets
