+++ tools/rum/facetsidebar.js
@@ -1,5 +1,6 @@
-import { pValue } from './cruncher.js';
-import { toHumanReadable, scoreCWV, escapeHTML } from './utils.js';
+import ListFacet from './list-facet.js';
+
+customElements.define('list-facet', ListFacet);
 
 const facetDecorators = {
   userAgent: {
@@ -13,6 +14,17 @@ const facetDecorators = {
   checkpoint: {
     label: 'Checkpoint',
     drilldown: 'flow.html',
+    labels: {
+      enter: 'Visit Entry',
+      loadresource: 'Fragment Loaded',
+      404: 'Not Found',
+      viewblock: 'Block Viewed',
+      viewmedia: 'Media Viewed',
+      click: 'Clicked',
+      error: 'JavaScript Error',
+      utm: 'Marketing Campaigns',
+      consent: 'Consent',
+    },
   },
   filter: {
     hidden: true,
@@ -115,67 +127,7 @@ export default class FacetSidebar {
     }
   }
 
-  computeTSV(facetName, mode, placeholders, show = {}, numOptions = 10) {
-    const facetEntries = this.dataChunks.facets[facetName];
-    const tsv = [];
-    const optionKeys = facetEntries.map((f) => f.value);
-    if (optionKeys.length) {
-      tsv.push(`${facetName}\tcount\tlcp\tcls\tinp`);
-      const filterKeys = facetName === 'checkpoint' && mode !== 'all';
-      const filteredKeys = filterKeys
-        ? optionKeys.filter((a) => !!(placeholders[a]))
-        : optionKeys;
-      const nbToShow = show[facetName] || numOptions;
-      facetEntries
-        .filter((entry) => !filterKeys || filteredKeys.includes(entry.value))
-        .slice(0, nbToShow)
-        .forEach((entry) => {
-          const CWVDISPLAYTHRESHOLD = 10;
-          let lcp = '-';
-          if (entry.metrics.lcp && entry.metrics.lcp.count >= CWVDISPLAYTHRESHOLD) {
-            const lcpValue = entry.metrics.lcp.percentile(75);
-            lcp = `${toHumanReadable(lcpValue / 1000)} s`;
-          }
-
-          let cls = '-';
-          if (entry.metrics.cls && entry.metrics.cls.count >= CWVDISPLAYTHRESHOLD) {
-            const clsValue = entry.metrics.cls.percentile(75);
-            cls = `${toHumanReadable(clsValue)}`;
-          }
-
-          let inp = '-';
-          if (entry.metrics.inp && entry.metrics.inp.count >= CWVDISPLAYTHRESHOLD) {
-            const inpValue = entry.metrics.inp.percentile(75);
-            inp = `${toHumanReadable(inpValue / 1000)} s`;
-          }
-
-          tsv.push(`${entry.name}\t${entry.value}\t${lcp}\t${cls}\t${inp}`);
-        });
-      tsv.push(...facetEntries
-        .filter((entry) => !filterKeys || filteredKeys.includes(entry.value))
-        .slice(0, nbToShow)
-        .map((entry) => `${entry.value}\t${entry.metrics.pageViews.sum}\t${entry.metrics.lcp.percentile(75)}\t${entry.metrics.cls.percentile(75)}\t${entry.metrics.inp.percentile(75)}`));
-    }
-    return tsv;
-  }
-
-  updateFacets(focus, mode, placeholders, show = {}) {
-    const createLabelHTML = (labelText, usePlaceholders) => {
-      if (labelText.startsWith('https://') && labelText.includes('media_')) {
-        return `<img src="${labelText}?width=750&format=webply&optimize=medium"">`;
-      }
-
-      if (labelText.startsWith('https://')) {
-        return `<a href="${labelText}" target="_new">${labelText}</a>`;
-      }
-
-      if (usePlaceholders && placeholders[labelText]) {
-        return (`${placeholders[labelText]} [${labelText}]`);
-      }
-      return escapeHTML(labelText);
-    };
-
-    const numOptions = mode === 'all' ? 20 : 10;
+  updateFacets(focus, mode) {
     const filterTags = document.querySelector('.filter-tags');
     filterTags.textContent = '';
     const addFilterTag = (name, value) => {
@@ -189,205 +141,35 @@ export default class FacetSidebar {
     if (this.elems.filterInput.value) addFilterTag('text', this.elems.filterInput.value);
     if (focus) addFilterTag(focus);
 
-    const url = new URL(window.location);
-
     this.elems.facetsElement.textContent = '';
     const keys = Object.keys(this.dataChunks.facets)
       // only show facets that have no decorators or are not hidden
       .filter((key) => !facetDecorators[key] || !facetDecorators[key].hidden);
     keys.forEach((facetName) => {
-      const facetEntries = this.dataChunks.facets[facetName];
-      const optionKeys = facetEntries.map((f) => f.value);
-      if (optionKeys.length) {
-        const fieldSet = document.createElement('fieldset');
-        fieldSet.classList.add(`facet-${facetName}`);
-        const legend = document.createElement('legend');
-        legend.textContent = facetDecorators[facetName]?.label || facetName;
-        const clipboard = document.createElement('span');
-        clipboard.className = 'clipboard';
-        legend.append(clipboard);
-        if (facetDecorators[facetName]?.drilldown && url.searchParams.get('drilldown') !== facetName) {
-          const drilldown = document.createElement('a');
-          drilldown.className = 'drilldown';
-          drilldown.href = facetDecorators[facetName].drilldown;
-          drilldown.title = 'Drill down to more details';
-          drilldown.textContent = '';
-          drilldown.addEventListener('click', () => {
-            const drilldownurl = new URL(drilldown.href, window.location);
-            drilldownurl.search = new URL(window.location).search;
-            drilldownurl.searchParams.delete(facetName);
-            drilldownurl.searchParams.set('drilldown', facetName);
-            drilldown.href = drilldownurl.href;
-          });
-          legend.append(drilldown);
-        } else if (url.searchParams.get('drilldown') === facetName) {
-          const drillup = document.createElement('a');
-          drillup.className = 'drillup';
-          drillup.href = facetDecorators[facetName].drilldown;
-          drillup.title = 'Return to previous level';
-          drillup.textContent = '';
-          drillup.addEventListener('click', () => {
-            const drillupurl = new URL('explorer.html', window.location);
-            drillupurl.search = new URL(window.location).search;
-            drillupurl.searchParams.delete(facetName);
-            drillupurl.searchParams.delete('drilldown', facetName);
-            drillup.href = drillupurl.href;
-          });
-          legend.append(drillup);
-        }
-
-        fieldSet.append(legend);
-        const filterKeys = facetName === 'checkpoint' && mode !== 'all';
-        const filteredKeys = filterKeys
-          ? optionKeys.filter((a) => !!(placeholders[a]))
-          : optionKeys;
-        const nbToShow = show[facetName] || numOptions;
-        facetEntries
-          .filter((entry) => !filterKeys || filteredKeys.includes(entry.value))
-          .slice(0, nbToShow)
-          .forEach((entry) => {
-            const div = document.createElement('div');
-            const input = document.createElement('input');
-            input.type = 'checkbox';
-            input.value = entry.value;
-            input.checked = url.searchParams.getAll(facetName).includes(entry.value);
-            if (input.checked) {
-              addFilterTag(facetName, entry.value);
-              div.ariaSelected = true;
-            }
-            input.id = `${facetName}=${entry.value}`;
-            div.addEventListener('click', (evt) => {
-              if (evt.target !== input) input.checked = !input.checked;
-              evt.stopPropagation();
-              this.dispatchEvent('change', this);
-            });
-
-            const label = document.createElement('label');
-            label.setAttribute('for', `${facetName}-${entry.value}`);
-            label.innerHTML = `${createLabelHTML(entry.value, facetName === 'checkpoint')} (${toHumanReadable(entry.metrics.pageViews.sum)})`;
-
-            const ul = document.createElement('ul');
-            ul.classList.add('cwv');
-
-            async function addSignificanceFlag(element, metric, baseline) {
-              const p = pValue(metric.values, baseline.values);
-              if (p < 0.05) {
-                element.classList.add('significant');
-              } else if (p < 0.1) {
-                element.classList.add('interesting');
-              }
-              element.dataset.pValue = p;
-            }
-
-            const CWVDISPLAYTHRESHOLD = 10;
-            // display core web vital to facets
-            // add lcp
-            let lcp = '-';
-            let lcpScore = '';
-            const lcpLI = document.createElement('li');
-            lcpLI.title = 'LCP';
-            if (entry.metrics.lcp && entry.metrics.lcp.count >= CWVDISPLAYTHRESHOLD) {
-              const lcpValue = entry.metrics.lcp.percentile(75);
-              lcp = `${toHumanReadable(lcpValue / 1000)} s`;
-              lcpScore = scoreCWV(lcpValue, 'lcp');
-              addSignificanceFlag(lcpLI, entry.metrics.lcp, this.dataChunks.totals.lcp);
-              lcpLI.title += ` - based on ${entry.metrics.lcp.count} samples`;
-            } else {
-              lcpLI.title += ` - not enough samples (${entry.metrics.lcp.count})`;
-            }
-            lcpLI.classList.add(`score-${lcpScore}`);
-            lcpLI.textContent = lcp;
-            ul.append(lcpLI);
-
-            // add cls
-            let cls = '-';
-            let clsScore = '';
-            const clsLI = document.createElement('li');
-            clsLI.title = 'CLS';
-            if (entry.metrics.cls && entry.metrics.cls.count >= CWVDISPLAYTHRESHOLD) {
-              const clsValue = entry.metrics.cls.percentile(75);
-              cls = `${toHumanReadable(clsValue)}`;
-              clsScore = scoreCWV(clsValue, 'cls');
-              addSignificanceFlag(clsLI, entry.metrics.cls, this.dataChunks.totals.cls);
-              clsLI.title += ` - based on ${entry.metrics.cls.count} samples`;
-            } else {
-              clsLI.title += ` - not enough samples (${entry.metrics.cls.count})`;
-            }
-            clsLI.classList.add(`score-${clsScore}`);
-            clsLI.textContent = cls;
-            ul.append(clsLI);
-
-            // add inp
-            let inp = '-';
-            let inpScore = '';
-            const inpLI = document.createElement('li');
-            inpLI.title = 'INP';
-            if (entry.metrics.inp && entry.metrics.inp.count >= CWVDISPLAYTHRESHOLD) {
-              const inpValue = entry.metrics.inp.percentile(75);
-              inp = `${toHumanReadable(inpValue / 1000)} s`;
-              inpScore = scoreCWV(inpValue, 'inp');
-              addSignificanceFlag(inpLI, entry.metrics.inp, this.dataChunks.totals.inp);
-              inpLI.title += ` - based on ${entry.metrics.inp.count} samples`;
-            } else {
-              inpLI.title += ` - not enough samples (${entry.metrics.inp.count})`;
-            }
-
-            inpLI.classList.add(`score-${inpScore}`);
-            inpLI.textContent = inp;
-            ul.append(inpLI);
-
-            div.append(input, label, ul);
-            fieldSet.append(div);
-          });
-
-        if (filteredKeys.length > nbToShow) {
-          // add "more" link
-          const div = document.createElement('div');
-          div.className = 'load-more';
-          const more = document.createElement('label');
-          more.textContent = 'more...';
-          more.addEventListener('click', (evt) => {
-            evt.preventDefault();
-            // increase number of keys shown
-            this.updateFacets(
-              focus,
-              mode,
-              placeholders,
-              { [facetName]: (show[facetName] || numOptions) + numOptions },
-            );
-          });
-
-          div.append(more);
-
-          const all = document.createElement('label');
-          all.textContent = `all (${filteredKeys.length})`;
-          all.addEventListener('click', (evt) => {
-            evt.preventDefault();
-            // increase number of keys shown
-            this.updateFacets(
-              focus,
-              mode,
-              placeholders,
-              { [facetName]: filteredKeys.length },
-            );
-          });
-          div.append(all);
-          const container = document.createElement('div');
-          container.classList.add('more-container');
-          container.append(div);
-          fieldSet.append(container);
-        }
-
-        legend.addEventListener('click', () => {
-          const tsv = this.computeTSV(facetName, mode, placeholders, show, numOptions);
-          navigator.clipboard.writeText(tsv.join('\r\n'));
-          const toast = document.getElementById('copied-toast');
-          toast.ariaHidden = false;
-          setTimeout(() => { toast.ariaHidden = true; }, 3000);
-        });
-
-        this.elems.facetsElement.append(fieldSet);
+      const facetEl = document.createElement(facetDecorators[facetName]?.component || 'list-facet');
+      facetEl.setAttribute('facet', facetName);
+      if (mode) facetEl.setAttribute('mode', mode);
+      if (focus) facetEl.setAttribute('focus', focus);
+      facetEl.setAttribute('drilldown', facetDecorators[facetName]?.drilldown);
+      facetEl.data = this.dataChunks;
+      facetEl.parent = this;
+      if (facetDecorators[facetName]?.label) {
+        const flabel = document.createElement('legend');
+        flabel.textContent = facetDecorators[facetName].label;
+        facetEl.append(flabel);
       }
+      if (facetDecorators[facetName]?.labels) {
+        const fnames = document.createElement('dl');
+        Object.entries(facetDecorators[facetName].labels).forEach(([key, value]) => {
+          const dt = document.createElement('dt');
+          dt.textContent = key;
+          const dd = document.createElement('dd');
+          dd.textContent = value;
+          fnames.append(dt, dd);
+        });
+        facetEl.append(fnames);
+      }
+      this.elems.facetsElement.append(facetEl);
     });
   }
 }
